ci(docker): add example environment file for EAS
This commit introduces a new example environment file `.eas.example.env` in the `docker` directory. The file includes a placeholder `EXPO_TOKEN` which is required for configuring EAS (Expo Application Services) in the project. This file serves as a template for setting up the necessary environment variables for EAS integration.